Fix togglint topbar if needed. Move everything in header to Appbar component
