+++ src/components/AppBar.js
@@ -1,10 +1,18 @@
 import React from 'react'
+import Toggle from 'material-ui/Toggle'
+import { Toolbar, ToolbarGroup } from 'material-ui/Toolbar'
 import TreeSearchBar from './TreeSearchBar'
 import StatusMessage from './StatusMessage'
-/**
- * A simple example of `AppBar` with an icon on the right.
- * By default, the left icon is a navigation-menu.
- */
+import GoogleButtons from './GoogleButtons'
+
+const styles = {
+  toggle: {
+    backgroundColor: '#5DBCD2'
+  },
+  labelStyle: {
+    color: '#FFF'
+  }
+}
 
 export default class AppBar extends React.Component {
   constructor(props) {
@@ -33,10 +41,29 @@ export default class AppBar extends React.Component {
         id="top-bar-header"
         style={{ background: '#5DBCD2', padding: 1 }}
       >
+        <div className="Header_root" id="header_root">
+          <Toolbar style={styles.toggle}>
+            <ToolbarGroup firstChild={true}>
+              <Toggle
+                label="Piilota / näytä aktiviteetit"
+                labelPosition="right"
+                style={styles.toggle}
+                onClick={() => this.props.toggleTopBar()}
+                labelStyle={styles.labelStyle}
+              />
+            </ToolbarGroup>
+            <ToolbarGroup>
+              <GoogleButtons />
+            </ToolbarGroup>
+          </Toolbar>
+        </div>
 
-      <StatusMessage />
-
-        <TreeSearchBar getHeight={this.getHeight} />
+        {this.props.headerVisible ? (
+          <div>
+            <StatusMessage />
+            <TreeSearchBar getHeight={this.getHeight} />
+          </div>
+        ) : null}
       </div>
     )
   }
